+++ src/System/LodgeService.java
@@ -16,7 +16,7 @@ public class LodgeService extends ServiceClass {
         costumers = new DoublyLinkedList<Student>();
     }
 
-    public long getPrice() { return monthly_cost; }
+    public double getPrice() { return monthly_cost; }
 
     public void newCostumer(Student student) {
         costumers.addLast(student);
